refactor(desktop): split app shell chat actions
