+++ apps/desktop/src/main/__tests__/renderer-shell-source-helpers.ts
@@ -12,6 +12,7 @@ const sourcePaths = [
   'app-shell-skill-actions.ts',
   'app-shell-project-actions.ts',
   'app-shell-open-skill-action.ts',
+  'app-shell-chat-actions.ts',
   'app-shell-import-actions.ts',
   'app-shell-stop-action.ts',
   'app-shell-copy.ts',
